backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pytest==8.0.0
httpx==0.26.0
python-dotenv==1.0.1
pydantic==2.6.1
supabase==2.4.5
ofxparse==0.21
pytest-asyncio==0.23.5
python-multipart==0.0.9 